+++ tools/wurf_dependency_bundle.py
@@ -7,57 +7,92 @@ compatible newer version of a specific library is available. The git
 tags must be named after the Semantic Versioning scheme defined here:
 www.semver.org
 
-The wscript will look like this:
-
-def options(opt):
-    opt.load('wurf_dependency_bundle')
-
-def configure(conf):
-    conf.load('wurf_dependency_bundle')
-
+This tool is loaded automatically in "wurf_common_tools".
 """
+
+import os
 
 from waflib.Configure import conf
 from waflib import Utils
 from waflib import Errors
+from waflib import ConfigSet
 
-import os
 
-OPTIONS_NAME = 'dependency options'
+OPTIONS_NAME = 'Dependency options'
 """ Name of the options group """
-
-DEFAULT_BUNDLE = 'ALL'
-""" Name of the default bundle """
 
 DEFAULT_BUNDLE_PATH = 'bundle_dependencies'
 """ Default folder to use for bundled dependencies """
 
-DEPENDENCY_PATH_KEY = '%s_DEPENDENCY_PATH'
-""" Destination of the dependency paths in the options """
+DEPENDENCY_PATH_KEY = '%s-path'
+""" Key to the dependency paths in the options """
 
-DEPENDENCY_CHECKOUT_KEY = '%s_DEPENDENCY_CHECKOUT'
-""" Destination of the dependency checkouts in the options """
+DEPENDENCY_CHECKOUT_KEY = '%s-use-checkout'
+""" Key to the dependency checkouts in the options """
 
 dependencies = dict()
-""" Dictionary storing the dependency information """
+""" Dictionary that stores the dependency resolvers """
+
+dependency_dict = dict()
+""" Dictionary to store and access the dependency paths by name """
+
+dependency_list = []
+""" List to store the dependency paths in the order of their definition """
 
 
-def add_dependency(opt, resolver):
+@conf
+def add_dependency(ctx, resolver, recursive_resolve=True):
     """
     Adds a dependency.
     :param resolver: a resolver object which is responsible for downloading
                      the dependency if necessary
+    :param recursive_resolve: specifies if it is allowed to recurse into the
+                     dependency wscript after the dependency is resolved
     """
     name = resolver.name
 
-    if name in dependencies:
+    if len(dependencies) == 0 and name != 'waf-tools':
+        ctx.fatal('waf-tools should be added before other dependencies')
 
+    if name in dependencies:
         if type(resolver) != type(dependencies[name]) or \
            dependencies[name] != resolver:
-            raise Errors.WafError('Incompatible dependency added %r <=> %r '
-                                  % (resolver, dependencies[name]))
-    else:
+            ctx.fatal('Incompatible dependency resolvers %r <=> %r '
+                       % (resolver, dependencies[name]))
+
+    # Skip dependencies that were already resolved
+    if name not in dependencies:
+
         dependencies[name] = resolver
+
+        bundle_opts = ctx.opt.get_option_group(OPTIONS_NAME)
+        add = bundle_opts.add_option
+
+        add('--%s-path' % name,
+            dest=DEPENDENCY_PATH_KEY % name,
+            default=False,
+            help='Path to %s' % name)
+
+        add('--%s-use-checkout' % name,
+            dest=DEPENDENCY_CHECKOUT_KEY % name,
+            default=False,
+            help='The checkout to use for %s' % name)
+
+        # The dependency resolvers are allowed to download dependencies
+        # if this is an "active" resolve step
+        if ctx.active_resolvers:
+            # Resolve this dependency immediately
+            path = resolve_dependency(ctx, name)
+            # Recurse into this dependency
+            if recursive_resolve:
+                ctx.recurse([path])
+        # Otherwise check if we already stored the path to this dependency
+        # when we resolved it (during an "active" resolve step)
+        elif name in ctx.env['DEPENDENCY_DICT']:
+            path = ctx.env['DEPENDENCY_DICT'][name]
+            # Recurse into this dependency
+            if recursive_resolve:
+                ctx.recurse([path])
 
 
 def expand_path(path):
@@ -81,22 +116,80 @@ def options(opt):
 
     add = bundle_opts.add_option
 
-    add('--bundle', default=DEFAULT_BUNDLE, dest='bundle',
-        help="Which dependencies to bundle")
-
     add('--bundle-path', default=DEFAULT_BUNDLE_PATH, dest='bundle_path',
-        help="The folder used for downloaded dependencies")
+        help='The folder where the bundled dependencies are downloaded. '
+             'Default folder: "{}"'.format(DEFAULT_BUNDLE_PATH))
 
-    for dependency in dependencies:
-        add('--%s-path' % dependency,
-            dest=DEPENDENCY_PATH_KEY % dependency,
-            default=False,
-            help='path to %s' % dependency)
 
-        add('--%s-use-checkout' % dependency,
-            dest=DEPENDENCY_CHECKOUT_KEY % dependency,
-            default=False,
-            help='The checkout to use for %s' % dependency)
+def resolve_dependency(ctx, name):
+
+    # If the user specified a path for this dependency
+    key = DEPENDENCY_PATH_KEY % name
+    dependency_path = getattr(ctx.options, key, None)
+
+    if dependency_path:
+
+        dependency_path = expand_path(dependency_path)
+
+        ctx.start_msg('User resolve dependency %s' % name)
+        ctx.end_msg(dependency_path)
+
+    else:
+        # Download the dependency to bundle_path
+
+        # Get the path where the bundled dependencies should be placed
+        bundle_path = expand_path(ctx.options.bundle_path)
+        Utils.check_dir(bundle_path)
+
+        ctx.start_msg('Resolve dependency %s' % name)
+
+        key = DEPENDENCY_CHECKOUT_KEY % name
+        dependency_checkout = getattr(ctx.options, key, None)
+
+        dependency_path = dependencies[name].resolve(
+            ctx=ctx,
+            path=bundle_path,
+            use_checkout=dependency_checkout)
+
+        ctx.end_msg(dependency_path)
+
+    ctx.env['DEPENDENCY_DICT'][name] = dependency_path
+    dependency_list.append(dependency_path)
+    return dependency_path
+
+
+def resolve(ctx):
+    """
+    The resolve function for the bundle dependency tool
+    :param ctx: the resolve context
+    """
+    if ctx.active_resolvers:
+        # wurf_dependency_resolve and wurf_git are only loaded if this is
+        # an "active" resolve step (i.e. if we are configuring the project)
+        ctx.load('wurf_dependency_resolve')
+        # Create a dictionary to store the resolved dependency paths by name
+        ctx.env['DEPENDENCY_DICT'] = dict()
+    else:
+        # Reload the environment from a previously completed resolve step
+        # if resolve.config.py exists in the build directory
+        try:
+            path = os.path.join(ctx.bldnode.abspath(), 'resolve.config.py')
+            ctx.env = ConfigSet.ConfigSet(path)
+        except EnvironmentError:
+            pass
+
+
+def post_resolve(ctx):
+    """
+    This function runs after the resolve step is completed
+    :param ctx: the resolve context
+    """
+    if ctx.active_resolvers:
+        # The dependency_dict will be needed in later steps
+        dependency_dict.update(ctx.env['DEPENDENCY_DICT'])
+        # Save the environment that was created during the active resolve step
+        path = os.path.join(ctx.bldnode.abspath(), 'resolve.config.py')
+        ctx.env.store(path)
 
 
 def configure(conf):
@@ -104,118 +197,22 @@ def configure(conf):
     The configure function for the bundle dependency tool
     :param conf: the configuration context
     """
-    conf.load('wurf_dependency_resolve')
-
-    # Get the path where the bundled dependencies should be
-    # placed
-    bundle_path = expand_path(conf.options.bundle_path)
-
-    # List all the dependencies to be bundled
-    bundle_list = expand_bundle(conf, conf.options.bundle)
-
-    # List all the dependencies with an explicit path
-    explicit_list = explicit_dependencies(conf.options)
-
-    # Make sure that no dependencies were both explicitly specified
-    # and specified as bundled
-    overlap = set(bundle_list).intersection(set(explicit_list))
-
-    if len(overlap) > 0:
-        conf.fatal("Overlapping dependencies %r" % overlap)
-
-    conf.env['BUNDLE_DEPENDENCIES'] = dict()
-
-    # Loop over all dependencies and fetch the ones
-    # specified in the bundle_list
-    for name in bundle_list:
-
-        Utils.check_dir(bundle_path)
-
-        conf.start_msg('Resolve dependency %s' % name)
-
-        key = DEPENDENCY_CHECKOUT_KEY % name
-        dependency_checkout = getattr(conf.options, key, None)
-
-        dependency_path = dependencies[name].resolve(
-            ctx=conf,
-            path=bundle_path,
-            use_checkout=dependency_checkout)
-
-        conf.end_msg(dependency_path)
-
-        conf.env['BUNDLE_DEPENDENCIES'][name] = dependency_path
-
-    for name in explicit_list:
-        key = DEPENDENCY_PATH_KEY % name
-        dependency_path = getattr(conf.options, key)
-        dependency_path = expand_path(dependency_path)
-
-        conf.start_msg('User resolve dependency %s' % name)
-        conf.env['BUNDLE_DEPENDENCIES'][name] = dependency_path
-        conf.end_msg(dependency_path)
+    # Copy the dependency dict and list (that were filled in the resolve step)
+    # to the persistent environment of the real ConfigurationContext
+    conf.env['DEPENDENCY_LIST'] = dependency_list
+    conf.env['DEPENDENCY_DICT'] = dependency_dict
+    # The dependencies will be enumerated in the same order as
+    # they were defined in the wscripts (waf-tools must be the first)
+    for path in conf.env['DEPENDENCY_LIST']:
+        conf.recurse([path])
 
 
-def expand_bundle(conf, arg):
-    """
-    Expands the bundle arg so that e.g. 'ALL,-gtest' becomes the
-    right set of dependencies
-    :param arg: list of bundle dependencies arguments
-    """
-    if not arg:
-        return []
-
-    arg = arg.split(',')
-
-    if 'NONE' in arg and 'ALL' in arg:
-        conf.fatal('Cannot specify both ALL and NONE as dependencies')
-
-    candidate_score = dict([(name, 0) for name in dependencies])
-
-    def check_candidate(c):
-        if c not in candidate_score:
-            conf.fatal('Cannot bundle %s, since it is not specified as a'
-                       ' dependency' % c)
-
-    for a in arg:
-
-        if a == 'ALL':
-            for candidate in candidate_score:
-                candidate_score[candidate] += 1
-            continue
-
-        if a == 'NONE':
-            continue
-
-        if a.startswith('-'):
-            a = a[1:]
-            check_candidate(a)
-            candidate_score[a] -= 1
-
-        else:
-            check_candidate(a)
-            candidate_score[a] += 1
-
-    candidates = [n for n in candidate_score if candidate_score[n] > 0]
-    return candidates
-
-
-def explicit_dependencies(options):
-    """
-    Extracts the names of the dependencies where an explicit
-    path have been that have been specified
-    :param options: the OptParser object where Waf stores the build options
-    """
-    explicit_list = []
-
-    for name in dependencies:
-
-        key = DEPENDENCY_PATH_KEY % name
-        path = getattr(options, key, None)
-
-        if path:
-            explicit_list.append(name)
-
-    return explicit_list
+def build(bld):
+    # The BuildContext reloads the environment of the ConfigurationContext,
+    # so the dependencies will be enumerated in the same order as
+    # in the configure step
+    for path in bld.env['DEPENDENCY_LIST']:
+        bld.recurse([path])
 
 
 @conf
@@ -223,11 +220,7 @@ def has_dependency_path(self, name):
     """
     Returns true if the dependency has been specified
     """
-
-    if name in self.env['BUNDLE_DEPENDENCIES']:
-        return True
-
-    return False
+    return name in self.env['DEPENDENCY_DICT']
 
 
 @conf
@@ -235,12 +228,4 @@ def dependency_path(self, name):
     """
     Returns the dependency path
     """
-    return self.env['BUNDLE_DEPENDENCIES'][name]
-
-
-@conf
-def is_toplevel(self):
-    """
-    Returns true if the current script is the top-level wscript
-    """
-    return self.srcnode == self.path
+    return self.env['DEPENDENCY_DICT'][name]